Add ~/.theia directory
When this directory does not exist and you mount a file into this directory (e.g. `-v $(pwd)/settings.json:/home/theia/.theia/settings.json`), the directory is owned by root and Theia is not allowed to write into this folder. With this commit the directory is created as theia user and it is safe to mount a file there.